+++ CHANGELOG.md
@@ -1,6 +1,28 @@
+## v1.3 [2018-08-04]
+
+  * New API version 2018-08-04
+  * new asset states: `published`, `reviewing`, `approved`, `rejected`, `archived`
+  * new `AssetStatistics` fields: versions, views, downloads
+  * asset version support
+    - new `Asset` fields: `version`, `locked`, `parentId`
+    - new fork type 'version-forks'
+    - new fields IN `AssetForkParams`: `version`, `locked`
+    - new fields in `AssetListParams`: `version`, `original`, `head`
+    - new fields in `LinkListParams`: `version`, `original`, `head`
+
+### Breaking API changes
+
+- replaced asset states `analyzing` and `transcoding` with more generic state `processing`
+- renamed asset field `version` (hash of metadata revision) to `revision`
+- added new asset field `version` (user-defined string, default HEAD)
+- renamed asset metadata type from `MetaVersion` to `MetaRevision`
+- renamed asset metadata id field from `versionId` to `revisionId`
+- renamed asset metadata (hash) field `version` to `revision`
+- renamed asset metadata query result list from `versions` to `revisions` (this is internal and should not be noticable)
+
 ## v1.0.2 [2018-07-04]
 
-  * New API version 2018-04-10
+  * New API version 2018-07-04
   * Smart stash support and stash description
   * Add commit messages to metadata updates
   * Add job state group `all` to list all jobs